+++ ff_data_server/data/credentials/oanda_credentials/download_list.toml
@@ -1,6 +1,6 @@
 # here we put symbols and base data types that we want the server to download data for, the server will keep the historical data up to date
 # The data will be 5 second quote bars, if we were to download all data it is over 100Gb from 2005 to 2024.
-# Oanda has no max download limit for historical data, so you can get as much as you like.
+# Oanda has no max download limit for historical data, so you can get_requests as much as you like.
 
 # Remove `#` to enable a symbol.
 
@@ -9,9 +9,9 @@
 
 #yyyy-mm-dd we can download only from a specific date forward to avoid getting all data,
 # start_date:the earliest date for OANDA data is 2005-01-01 yyyy-mm-dd
-# If you make the date an earlier time, the server will get extra the data on start up, or on the next update cycle (if you don't want to restart the server).
+# If you make the date an earlier time, the server will get_requests extra the data on start up, or on the next update cycle (if you don't want to restart the server).
 # There is no way to download gaps in data, so if you are moving the start date backwards and you already have data, you should let the download finish before shutting down the server, or you will have to manually remove the data
-# And adjust the start date to the end of the last download to get the data from the gap you created, i will try to make an efficient self healing fn later.
+# And adjust the start date to the end of the last download to get_requests the data from the gap you created, i will try to make an efficient self healing fn later.
 symbols = [
     { symbol_name = "NAS100-USD", base_data_type = "QuoteBars", resolution = "5-S", start_date = "2005-01-01" },
     #{ symbol_name = "NAS100-USD", base_data_type = "QuoteBars", resolution = "1-M", start_date = "2024-01-01" },
