+++ ff_data_server/data/credentials/rithmic_credentials/download_list.toml
@@ -9,18 +9,18 @@
 # Candles,
 
 # Tick history for 1 symbol is over 60GB from 2019 to Current, this is per symbol! Rithmic Limits downloads to 40Gb per month per user.
-# If you have a cheap prop firm account it might pay to use it to get the initial data and save your real rithmic account for live updates.
+# If you have a cheap prop firm account it might pay to use it to get_requests the initial data and save your real rithmic account for live updates.
 # Keep in mind that if you have a prop account and a Rithmic01 or Rithmic4Colo account in the active folder, then rithmic will use the real accounts for data, be sure to remove the real accounts from the active folder if you want to use the prop account for data.
 
 #yyyy-mm-dd we can download only from a specific date forward to avoid getting all data,
 # start_date: the earliest date for Rithmic Tick data is 20019-06-02 yyyy-mm-dd (there is other data to 2011 but i don't support it yet for historical)
-# If you make the date an earlier time, the server will get extra the data on start up, or on the next update cycle (if you don't want to restart the server).
+# If you make the date an earlier time, the server will get_requests extra the data on start up, or on the next update cycle (if you don't want to restart the server).
 # There is no way to download gaps in data, so if you are moving the start date backwards and you already have data, you should let the download finish before shutting down the server, or you will have to manually remove the data
-# And adjust the start date to the end of the last download to get the data from the gap you created, I will try to make an efficient self-healing fn later.
+# And adjust the start date to the end of the last download to get_requests the data from the gap you created, I will try to make an efficient self-healing fn later.
 
 # Remove `#` to enable a symbol.
 
-# If you don't get data, you might need to move the start date forewards, data availability depends on many things. The server is designed to stop polling for data after x number of empty responses
+# If you don't get_requests data, you might need to move the start date forewards, data availability depends on many things. The server is designed to stop polling for data after x number of empty responses
 symbols = [
     { symbol_name = "MNQ", base_data_type = "Ticks", resolution = "1-T", start_date = "2019-06-02"}, # start_date: start date is optional, the earliest date for Rithmic tick is 2019-06-02 yyyy-mm-dd
     { symbol_name = "MNQ", base_data_type = "Candles", resolution = "1-M", start_date = "2021-01-01"},
