+++ ff_data_server/docker builds/apple_silicone_linux/Dockerfile
@@ -16,7 +16,7 @@ RUN cargo build --release
 FROM --platform=linux/arm64 debian:bookworm-slim
 
 # Install runtime dependencies necessary for your application
-RUN apt-get update && apt-get install -y \
+RUN apt-get_requests update && apt-get_requests install -y \
     ca-certificates \
     openssl \
     && rm -rf /var/lib/apt/lists/*
