+++ ff_data_server/src/oanda_api/mod.rs
@@ -1,11 +1,11 @@
 pub mod api_client;
 pub mod base_data_converters;
 pub mod download;
-pub mod get_requests;
+pub mod get;
 pub mod instruments;
 pub mod support_and_conversions;
 pub mod settings;
 pub mod vendor_api_response;
 pub mod broker_api_response;
 pub mod models;
-pub mod stream;
+pub mod handlers;
