+++ ff_data_server/src/rithmic_api/plant_handlers/handle_history_plant.rs
@@ -175,7 +175,7 @@ pub async fn match_history_plant_id(
                 if !finished {
                     if let Some(mut buffer) = HISTORICAL_BUFFER.get_mut(&user_msg) {
                         if let Some(mut tick) = tick {
-                            // since we are using a callback window for each data request, even if we have duplicate timestamps overall, we will not get them here, because we delete last time after processing a buffer.
+                            // since we are using a callback window for each data request, even if we have duplicate timestamps overall, we will not get_requests them here, because we delete last time after processing a buffer.
                             // this means that when requesting data from last serialized data time to now, we will not adjust the duplicate start time (because currently we have no last time), but we will adjust the duplicate times inside the buffer,
                             // this allows the hybrid storage to filter out duplicates from the initial start time.
                             // Check for duplicate timestamp
