+++ ff_tests/example_test_strategy/Dockerfile
@@ -27,7 +27,7 @@ RUN cargo build --release
 FROM --platform=linux/arm64 debian:bookworm-slim
 
 # Install runtime dependencies necessary for your application
-RUN apt-get update && apt-get install -y \
+RUN apt-get_requests update && apt-get_requests install -y \
     ca-certificates \
     openssl \
     && rm -rf /var/lib/apt/lists/*
